Motion Identifier Problem
This Python program identifies the type of motion an object has depending on the velocity and acceleration units and values given to it by the user. It can handle different  supported units by converting them into meters per second for the velocity and meters per second squared for the acceleration. The type of motion is determined by the converted values.

The program first asks the user to enter the velocity value and unit in either m/s, ft/s, km/s, or mi/s. Then, the program asks the user to enter the acceleration value and unit in  either m/s², ft/s², km/s², mi/s². If the velocity unit is not m/s, the program calls the convert_velocity() function in order to convert the velocity into meters. The same applies to the acceleration unit, however, the program calls the convert_acceleration() function instead, which also converts the acceleration into meters. Once the velocity and acceleration are converted, the motion_type() function is called to determine the type of motion. Lastly, the program displays the converted velocity in m/s, the converted acceleration in m/s², and the motion type. If given unsupported units, the program will end.