motion_identifier.py
This python program asks the user to enter the velocity value and unit in m/s, ft/s, km/s, or mi/s, and the acceleration value and unit in m/s², ft/s², km/s², or mi/s². The program then calls the convert_velocity() function and if the unit is not in meters, it is converted. The same applies for acceleration but the convert_acceleration() function is called instead. The program then calls the motion_type() function, where the converted values are then used to determine the type of motion. Finally, the program displays the converted velocity and acceleration along with the type of motion.

The program quits when given an unsupported unit.